+++ src/visualize/+mag/+graphics/+chart/Line.m
@@ -5,11 +5,15 @@ classdef Line < mag.graphics.chart.Chart & mag.graphics.mixin.ColorSupport
         % AXIS Axis along which to plot.
         Axis (1, 1) string {mustBeMember(Axis, ["x", "y"])} = "y"
         % VALUE Line value.
-        Value (1, 1)
+        Value (1, :)
         % STYLE Line style.
         Style (1, 1) string = "-"
         % LABEL Line label.
-        Label string {mustBeScalarOrEmpty}
+        Label (1, :) string = string.empty()
+        % HORIZONTALALIGNMENT Horizontal alignment of label.
+        HorizontalAlignment (1, 1) string {mustBeMember(HorizontalAlignment, ["left", "center", "right"])} = "right"
+        % VERTICALALIGNMENT Vertical alignment of label.
+        VerticalAlignment (1, 1) string {mustBeMember(VerticalAlignment, ["top", "middle", "bottol"])} = "top"
     end
 
     methods
@@ -26,7 +30,7 @@ classdef Line < mag.graphics.chart.Chart & mag.graphics.mixin.ColorSupport
         function graph = plot(this, ~, axes, ~)
 
             arguments (Input)
-                this
+                this (1, 1) mag.graphics.chart.Line
                 ~
                 axes (1, 1) matlab.graphics.axis.Axes
                 ~
@@ -36,11 +40,14 @@ classdef Line < mag.graphics.chart.Chart & mag.graphics.mixin.ColorSupport
                 graph (1, :) matlab.graphics.Graphics
             end
 
+            args = {axes, this.Value, this.Style, this.Label, ...
+                "LabelHorizontalAlignment", this.HorizontalAlignment, "LabelVerticalAlignment", this.VerticalAlignment};
+
             switch this.Axis
                 case "x"
-                    graph = xline(axes, this.Value, this.Style, this.Label);
+                    graph = xline(args{:});
                 case "y"
-                    graph = yline(axes, this.Value, this.Style, this.Label);
+                    graph = yline(args{:});
             end
 
             this.applyColorStyle(graph);